hattucci/server/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
passlib==1.7.4
Werkzeug==3.0.4
reportlab==4.2.2